`Added PingTimeout method and corresponding tests`
` `

`Implemented a new method, PingTimeout, in the NetworkService class that takes two integers as input and returns their sum. Also added a new test method, NetworkService_PingTimeout_ReturnSum, in the NetworkUtility.Tests.PingTest namespace to validate the functionality of the PingTimeout method. This test method uses the Theory and InlineData attributes to test multiple sets of data.`